requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
razorpay==1.4.1
